abscissa_derive: custom derive for Command
Replaces the previous macro_rules!
